fixing bug

+++ Fusheng_Wang_image_bio/src/main/java/org/pulasthi/dsctools/DistanceCalculation.java
@@ -112,8 +112,8 @@ public class DistanceCalculation {
                 }
                 byteBuffer.clear();
                 byteBuffer.asShortBuffer().put(row);
-                System.out.println("********************* :" + (filePosition + i*numPoints*2) + "********************\n");
-                fc.write(byteBuffer,(filePosition + i*numPoints*2));
+                System.out.println("********************* :" + (filePosition + ((long)i)*numPoints*2) + "********************\n");
+                fc.write(byteBuffer,(filePosition + ((long)i)*numPoints*2));
             }
 
             fc.close();
